Repository pattern
* Repository Pattern

* Add Validation to Users endpoint

* Remove Entities namespace

* Add Contracts namespace

* Fix Repositories

* Fix namespace

* Rename UsersController to UserController

* Fix namespace

* Move interfaces to app\Contracts

* Add DB dump

* Methods should be compatible with interface

* Set up tokens and scopes for tests

* Rename token

* Update doc blocks

* Fix namespace in seeders

* Use DbHelper to populate and clean up database

* Fix namespace in Auth Controller

* Add Get all users tests

* Fix Namespace in User Controller

* Add Codeception tests for create, list, update, delete a user

* Rename UserFormRequest to UserRequest

* Extend Dingo\Api\Http\FormRequest

* Add all CRUD tests and update all CRUD methods for Users Controller

* Add Password Hashing

* Add user and client resolvers

* Fix namespace on the Auth config file and update resolvers

* Add user authentication

* Add user token for Password grant and test data

* Don't use bcrypt since setPasswordAttribute already uses Hash::make

* Add password grant test tokens

* Fix a typo for admintoken

* Use repository instead of model

* Ensure that passwords are 8 characters or greater

* Fix user and admin tokens

* Ensure that user details can only be modified by owner

* Fix typo in User Cest

* Rename scenario

* Use TRUNCATE instead of DELETE to clear user's table

* Fix assignment

* Add role seeder and get role method for isAdmin

* Add admin authentication

* Add roles and OAuth test data

* Fix tests

* Add foreign keys in pivot table

* Add IsAdmin()

* Add getRoles()

* Fix relationships

* split Requests for easier authorization checks

* Change datatypes for groups,orgs and settings

* Remove an unnecessary method - getUserDetails

* Add tests that check on wrong type of authentication

* Use spaces instead of tabs for indentation

* Add doc blocks and make test cases readable

* update test case descriptions

+++ database/migrations/2015_10_08_103035_create_organization_groups_table.php
@@ -6,37 +6,37 @@ use Illuminate\Database\Schema\Blueprint;
 class CreateOrganizationGroupsTable extends Migration
 {
 
-	/**
-	 * Run the migrations.
-	 *
-	 * @return void
-	 */
-	public function up()
-	{
-		Schema::create('organization_groups', function ($table) {
-			$table->engine = 'InnoDB';
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('organization_groups', function ($table) {
+            $table->engine = 'InnoDB';
 
-			$table->bigInteger('organization_id')->unsigned()->default(0);
-			$table->bigInteger('group_id')->unsigned();
+            $table->integer('organization_id')->unsigned()->default(0);
+            $table->integer('group_id')->unsigned();
 
-			$table->foreign('organization_id')
-				->references('id')->on('organizations')
-				->onDelete('cascade');
+            $table->foreign('organization_id')
+                ->references('id')->on('organizations')
+                ->onDelete('cascade');
 
-			$table->foreign('group_id')
-				->references('id')->on('groups')
-				->onDelete('cascade');
-		});
-	}
+            $table->foreign('group_id')
+                ->references('id')->on('groups')
+                ->onDelete('cascade');
+        });
+    }
 
-	/**
-	 * Reverse the migrations.
-	 *
-	 * @return void
-	 */
-	public function down()
-	{
-		Schema::drop('organization_groups');
-	}
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::drop('organization_groups');
+    }
 
 }
