+++ database/migrations/2015_10_15_095401_create_group_users_table.php
@@ -6,28 +6,28 @@ use Illuminate\Database\Schema\Blueprint;
 class CreateGroupUsersTable extends Migration
 {
 
-	/**
-	 * Run the migrations.
-	 *
-	 * @return void
-	 */
-	public function up()
-	{
-		Schema::create('group_users', function (Blueprint $table) {
-			$table->bigInteger('group_id')->unsigned()->default(0);
-			$table->bigInteger('user_id')->unsigned();
-			$table->primary(['group_id', 'user_id']);
-		});
-	}
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('group_users', function (Blueprint $table) {
+            $table->integer('group_id')->unsigned()->default(0);
+            $table->integer('user_id')->unsigned();
+            $table->primary(['group_id', 'user_id']);
+        });
+    }
 
-	/**
-	 * Reverse the migrations.
-	 *
-	 * @return void
-	 */
-	public function down()
-	{
-		Schema::drop('group_users');
-	}
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::drop('group_users');
+    }
 
 }
